Remove the permutation of position in get_random_node_with_token
This is no longer necessary, because the graph brain model ensures
that every token is always present at token0_id in some node.

This simplifies the SQL in get_random_node_with_token and allows
us to remove the single-column tokenN_id indexes on the nodes table,
for a 16% brain size savings.