Prewarm the kernel buffer cache on start
This is a major performance improvement on a large brain, bringing
me from ~10 replies in half a second to well over 200 on average.

Prewarming the cache helps a great deal because of a combination
of factors. Brain data that is tightly connected is unlikely to
have good locality in the database, and the random walk for
candidate reply generation means that any database page could be
accessed during a reply.

Generating more candidate replies would help warm the cache, but
since that's a timed operation, a cold cache means fewer replies
and fewer pages get cached.