Yield all the permutations of search results in _generate_replies
The old behavior was to choose a pivot node, generate the end
and beginning of a sentence from that node, and then emit the
combination:

[loop 0] <begin 0> node <end 0>
[loop 1] <begin 1> node <end 1>
[loop 2] <begin 2> node <end 2>

But node is the potential center of any combination of begin and end
nodes. The new behavior is to emit all of those combinations:

[loop 0] <begin 0> node <end 0>
[loop 1] <begin 1> node <end 1>
         <begin 0> node <end 1>
         <begin 1> node <end 0>
[loop 2] <begin 2> node <end 2>
         <begin 0> node <end 2>
         <begin 1> node <end 2>
         <begin 2> node <end 0>
         <begin 2> node <end 1>

This allows us to create many more potential replies in our time slice
without incurring any more database load.