Use an n-gram's order (3, 2, 1, etc) as its key prefix
All n-gram counts were previously prefixed with "c". Here we
switch to prefixing them with their order:

	1token1
	2token1token2
	3token1token2token3

This reduces data locality, so we can't get e.g. a bigram and
all its following trigrams with one range query, but it's a
decent optimization when determing the n of an n-gram while
saving.